rename components/index to components/register
