+++ src/app/main.js
@@ -4,7 +4,7 @@ import Vue from 'vue'
 import VueRouter from 'vue-router'
 import App from './App.vue'
 
-import './components'
+import './components/register'
 
 Vue.use(VueRouter)
